+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.37.1] - 2024-06-03
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Dialog message component.
+
+### Removed
+- [Marek Kurańda](https://github.com/mjkuranda): Annotation with contact count.
+
 ## [0.37.0] - 2024-06-03
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Contact info list at `/contact-us`.